model2/requirements.txt
absl-py==0.10.0
alabaster==0.7.12
alembic==1.0.9
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appdirs==1.4.4
asn1crypto==1.0.1
asr-evaluation==2.0.4
astor==0.8.1
astroid==2.3.1
astropy==3.2.2
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blis==0.7.4
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.1.1
catalogue==2.0.4
certifi==2021.10.8
cffi==1.12.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
community==1.0.0b1
conda==4.11.0
conda-build==3.18.9
conda-package-handling==1.7.0
conda-verify==3.4.2
configparser==5.0.2
contextlib2==0.6.0
crayons==0.4.0
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.5
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
databricks-cli==0.9.1
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.1
distributed==2.5.2
docker==5.0.0
docker-pycreds==0.4.0
docutils==0.15.2
edit-distance==1.0.4
elmoformanylangs==0.0.4.post2
en-core-web-sm==3.0.0
entrypoints==0.3
et-xmlfile==1.0.1
eulxml==1.1.3
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
flatbuffers==2.0
fonttools==4.29.1
fsspec==0.5.2
future==0.18.2
gast==0.2.2
gensim==3.8.3
gevent==1.4.0
ghp-import==2.0.2
gitdb==4.0.7
GitPython==3.1.14
glob2==0.7
gmpy2==2.0.8
google-api-core==1.20.1
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==0.4.15
grpcio==1.31.0
gunicorn==20.1.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.18.1
hypothesis==4.14.0
idna==2.6
imageio==2.6.0
imagesize==1.1.0
img2vec-pytorch==1.0.1
importlib-metadata==4.11.3
influxdb==5.3.1
ipykernel==5.1.2
ipython==7.8.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itu.algs4==0.2.3
jdcal==1.4.1
jedi==0.15.1
jeepney==0.4.1
Jinja2==3.0.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.0.0
jupyter-core==4.9.2
jupyterlab==1.1.4
jupyterlab-pygments==0.1.2
jupyterlab-server==1.0.6
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
langid==1.1.6
lazy-object-proxy==1.4.2
libarchive-c==2.8
libclang==13.0.0
lief==0.9.0
line-profiler==2.1.2
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Mako==1.1.4
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
mccabe==0.6.1
memory-profiler==0.55.0
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.2.3
mkdocs-material==8.2.5
mkdocs-material-extensions==1.0.3
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mknotebooks==0.7.1
mock==3.0.5
more-itertools==8.12.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
murmurhash==1.0.5
navigator-updater==0.2.1
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.20.1
numpydoc==0.9.1
nxviz==0.7.3
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
openai==0.6.4
opencv-python==4.5.5.62
openpyxl==3.0.0
opt-einsum==3.3.0
overrides==3.1.0
packaging==20.9
palettable==3.3.0
pandas==1.3.5
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
pathy==0.5.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pip==22.0.3
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
powerlaw==1.4.6
pprofile==2.0.2
praw==7.5.0
prawcore==2.3.0
preshed==3.0.5
prometheus-client==0.7.1
prometheus-flask-exporter==0.18.1
prompt-toolkit==2.0.10
protobuf==3.14.0
psutil==5.6.3
ptyprocess==0.6.0
PuLP==2.6.0
py==1.8.0
py-heat==0.0.6
py-heat-magic==0.0.2
py4j==0.10.9
pyarrow==3.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydantic==1.7.3
pydicom==2.2.2
pyflakes==2.1.1
Pygments==2.11.2
pylint==2.4.2
pymdown-extensions==9.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pypng==0.0.21
pyrsistent==0.15.4
PySocks==1.7.1
pyspark==3.1.1
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-editor==1.0.4
python-louvain==0.16
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1
pyyaml_env_tag==0.1
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
querystring-parser==1.2.4
requests==2.23.0
requests-oauthlib==0.8.0
rope==0.14.0
rsa==3.4.2
ruamel_yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.6.2
seaborn==0.11.2
SecretStorage==3.1.1
selenium==3.141.0
Send2Trash==1.5.0
setuptools==50.3.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==3.0.0
smmap==3.0.5
snakeviz==2.0.1
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
spacy==3.0.6
spacy-legacy==3.0.5
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-fulltoc==1.2.0
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spotipy==2.16.1
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
sqlparse==0.4.1
srsly==2.4.1
statsmodels==0.10.1
sympy==1.4
tables==3.4.4
tabulate==0.8.9
tblib==1.4.0
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
texttable==1.6.4
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.3
thresholdclustering==1.1
toolz==0.10.0
torch==1.7.1
torchtext==0.9.1
torchvision==0.8.2
tornado==6.0.3
tqdm==4.60.0
traitlets==5.1.1
typer==0.3.2
typing-extensions==3.7.4.1
unicodecsv==0.14.1
update-checker==0.18.0
uritemplate==3.0.0
urllib3==1.25.9
virtualenv==20.4.3
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.1.7
webdriver-manager==3.4.1
webencodings==0.5.1
websocket-client==0.57.0
wer==0.1.0
Werkzeug==0.16.0
wheel==0.33.6
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
